genai/requirements_genai.txt
# GenAI Intelligence Layer Requirements
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0